# Faster iterating over (almost) empty `Dict` or `Set`