It is possible to utilize the existing code https://github.com/JuliaLang/julia/commit/0e127232a0fa974dfe29660dc8e6687cb683f964 defining `idxfloor` to improve performance of iterating over (almost) empty `Dict`/`Set`.

However, this PR cannot solve asymptotic complexity because the current implementation cannot shrink `Dict` so far (there is a todo for that). https://github.com/JuliaLang/julia/blob/b10aa56b501dd9917921f05d3240cfdee40dd63e/base/dict.jl#L238-L243.
 
Master
``` julia
julia> using BenchmarkTools

julia> d = Set{Int}();
julia> @btime minimum(d, init = 0);
  20.737 ns (0 allocations: 0 bytes)

julia> d = Set([1]);
julia> @btime minimum(d, init = 0);
  21.904 ns (0 allocations: 0 bytes)

julia> d = Set(1:10_000); empty!(d);
julia> @btime minimum(d, init = 0);
  6.581 μs (0 allocations: 0 bytes)

julia> d = Set(1:10_000); empty!(d); push!(d,1);
julia> @btime minimum(d, init = 0);
  6.665 μs (0 allocations: 0 bytes)
```

PR
``` julia
julia> using BenchmarkTools

julia> d = Set{Int}();
julia> @btime minimum(d, init = 0);
  14.968 ns (0 allocations: 0 bytes)

julia> d = Set([1]);
julia> @btime minimum(d, init = 0);
  16.615 ns (0 allocations: 0 bytes)

julia> d = Set(1:10_000); empty!(d);
julia> @btime minimum(d, init = 0);
  14.933 ns (0 allocations: 0 bytes)

julia> d = Set(1:10_000); empty!(d); push!(d,1);
julia> @btime minimum(d, init = 0);
  821.388 ns (0 allocations: 0 bytes)
```